// site/src/lib/generated/linguini/shared.d.ts
export type __lgl_name_6D61696E2E4672756974 = "apple" | "pear" | "orange";

export type __lgl_name_6D61696E2E53697A65 = "small" | "big";

export type __lgl_name_6D61696E2E4D6F6E6579 = number;

export type __lgl_name_6D61696E2E53686F727444617465 = Date | number | string;

export type __lgl_name_6D61696E2E4D6561737572656D656E74 = number;

export declare function selectBranch(
  key: string,
  branches: Record<string, string>,
): string;